+++ src/h2/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 """
-h2
+hyper-h2
 ~~
 
 A HTTP/2 implementation.
 """
-__version__ = '3.1.1dev0'
+__version__ = '4.0.0+dev'