+++ src/h2/frame_buffer.py
@@ -26,7 +26,7 @@ from .exceptions import (
 CONTINUATION_BACKLOG = 64
 
 
-class FrameBuffer(object):
+class FrameBuffer:
     """
     This is a data structure that expects to act as a buffer for HTTP/2 data
     that allows iteraton in terms of H2 frames.
@@ -130,7 +130,7 @@ class FrameBuffer(object):
     def __iter__(self):
         return self
 
-    def next(self):  # Python 2
+    def __next__(self):
         # First, check that we have enough data to successfully parse the
         # next frame header. If not, bail. Otherwise, parse it.
         if len(self.data) < 9:
@@ -149,13 +149,11 @@ class FrameBuffer(object):
         # Confirm the frame has an appropriate length.
         self._validate_frame_length(length)
 
-        # Don't try to parse the body if we didn't get a frame we know about:
-        # there's nothing we can do with it anyway.
-        if f is not None:
-            try:
-                f.parse_body(memoryview(self.data[9:9+length]))
-            except InvalidFrameError:
-                raise FrameDataMissingError("Frame data missing or invalid")
+        # Try to parse the frame body
+        try:
+            f.parse_body(memoryview(self.data[9:9+length]))
+        except InvalidFrameError:
+            raise FrameDataMissingError("Frame data missing or invalid")
 
         # At this point, as we know we'll use or discard the entire frame, we
         # can update the data.
@@ -169,7 +167,4 @@ class FrameBuffer(object):
         # frame in the sequence instead. Recurse back into ourselves to do
         # that. This is safe because the amount of work we have to do here is
         # strictly bounded by the length of the buffer.
-        return f if f is not None else self.next()
-
-    def __next__(self):  # Python 3
-        return self.next()
+        return f if f is not None else self.__next__()